+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fixed selecting map location when group marker is selected [#4363](https://github.com/rokwire/illinois-app/issues/4363).
+
 ## [6.1.11] - 2024-10-09
 ### Added
 - Extended error handling for SafeWalks SMS sending [#4363](https://github.com/rokwire/illinois-app/issues/4363).